Changed Licensing to LGPL and moved CalcLocalKMatrix_Quad to a seperate file - not working so fsr.
